News_agent/pyproject.toml
[project]
name = "reddit_post_check"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apscheduler>=3.11.0",
    "asyncpraw>=7.8.1",
    "dotenv>=0.9.9",
    "gradio>=5.49.1",
    "httpx>=0.28.1",
    "l>=0.11.0",
    "logfire>=4.13.2",
    "praw>=7.8.1",
    "pydantic-ai>=1.1.0",
    "pydantic-ai-slim[openai]>=1.0.18",
    "python-dotenv>=1.1.1",
    "python-telegram-bot>=22.5",
]
